requirements.txt
streamlit==1.32.0
tensorflow==2.16.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
matplotlib==3.8.3
pandas==2.2.1